Fix the RTC incrementing procedure
Previously at every instant in the simulation the time was wrongly
based on the current RTC increment value, as if it was configured
at the RTC since the beginning. Now the current RTC increment value
is used to compute solely the elapsed time since the previous RTC
update. This way, the entire history of RTC increment values
determines the RTC time scale.