Fix frequency offset estimation considering large SYNC intervals
Wrong behavior was resulting previously due the fact that the
frequency offset estimation was not prepared for SYNC intervals
larger than one second, which can occur when the SYNCs are only
processed after a "packet" (time offset) selection.

Add also a threshold for the freq offset correction and improve
the priting.