Add time offset computations
Introduce time offset measurement on the delay request-response processing. Add
also measurement of the true time offset between the RTC.

Furthermore, update the behavior of the subtraction of timestamps such that
negative seconds are allowed. This is essential for the time offset computations
to work.

Improve also the logging format such that columns can be more easily inspected.