Add modeling of RTC frequency offset
Include parameter on RTC class to define the range of a uniformly distributed
initial fractional frequency offset of the driving clock. Then, change the RTC
implementation such that the instantaneous frequency offset is always
considered. Add unit tests to make sure time-keeping is functional after
modification to account for the time-varying frequency offset.