Add CPU-efficient sample-mean implementations
When the sample mean is applied over windows of measurements that are displaced
by one index only, the result is effectively the same as a moving average. In
this case, more CPU-efficient moving average implementations can be used.

This patch adds two typical implementations for moving average. The first is a
recursive implementation of a conventional moving average. This is effectively a
FIR filter. The second is an exponentially weighted moving average (EWMA), which
is in principle an IIR filter, but with a very simple recursive implementation.