Fix final fit of LS efficient and results index
In order to make the efficient LS compatible with the other LS implementations,
take the final time offset fit at the corresponding instant, that is, the last
instant of the window, which is $t = (N-1)T$. Also, save the results on the same
indexes that the other LS implementations do.

With the changes of this patch, the results between all the three LS
implementations should match very closely.