Control freq offset estimation window in samples
Instead of determining the period in nanoseconds to take frequency offset
estimations, control in terms of samples. More importantly, do so in a way that,
even if the delta (in samples) is greater than 1, every sample produces a
frequency offset estimation. That is, for delta = 2, sample 2 estimates based on
sample 2 and sample 0, sample 3 estimates based on sample 3 and sample 1, and so
on. In other words, use overlapping observation windows.