Consider reversed master-slave direction on reader
We will use mostly the PDelay req-resp from the slave side when acquiring
timestamps. Hence, the master-slave order will be reversed. This order matters
when we compare the true time offset measurement (slave - master) with the
estimated time offset.

This patch changes the order of timestamps saved by the reader module such that
the sign of time and frequency offset computations still reflect difference
relative to master (slave minus master).