Rename "maxTE vs window" to "window optimizer"
Rename the demo and the associated module from the PTP package. Use the shorter
name for the module ("window.py") and name the process as "window
optimization". We may decide to optimize the window length based on another
metric (other than max|TE|) in the future. We may consider for example both
max|TE| and MTIE. Hence, "window optimizer" is a more generic future-proof
naming.