Process timestamp differences on sample-median
It makes more sense to try cancelling the median of t2-t1 differences with
respect to the median of t4-t3 differences. The previous approach had really no
mathematical reasoning, it was more about heuristics. With the new approach, it
will work as long as forward and reverse delays present similar
distributions. If not, maybe the asymmetry can be calibrated.