Add vectorized packet selection implementations
Add vectorized implementations for sample-average, sample-min, sample-max,
sample-median and sample-mode. Instead of processing each observation window
individually, stack all observation windows as lines of a matrix and process the
entire matrix at once. The results produced by all operators (except
sample-mode) are equivalent to the non-vectorized results, but with a dramatic
increase in computational speed.

The sample-mode implementation differs due to how the bin adjustment is
implemented. On the non-vectorized version, after each observation window, we
decide whether or not to adjust the bin width. However, if we do adjust it, the
previous estimates won't be re-computed. In contrast, on the vectorized version,
we choose to tune the bin based on a smaller matrix containing solely the
initial 100 observation windows. After the bin width is tuned, the entire matrix
is re-processed with the chosen bin width.

For comparison, the non-vectorized implementation was kept in the code, but it
may be removed in the future. To control whether or not to use the vectorized
implementation, an argument (vectorize) was added to the PktSelection.process()
function. The vectorized implementation is set to be used by default.