Make coarse-fine window optimizer passes optional
The coarse-fine pass strategy is useful when running the window optimizer to
simply find the best window lengths. It leads to fast results. Nevertheless,
when the goal is to generate a plot of the max|TE| behavior vs. window length,
e.g. for a paper or presentation, it is better to be able to evaluate all
possible window lengths on a range, so that we get a roughly continuous convex
curve.

This patch adds this option to disable the coarse-fine pass strategy and instead
evaluate all possible values on a range.