Fix paths used by window optimizer
The way that they were defined previously would only work in case the window
optimization module was launched from the root directory of the project. In
contrast, it would fail when launching from a notebook located at
"rootdir/notebooks/". This patch fixes this issue by relying on a reliable
reference to the root directory.

Update also the reader demo notebook such that it can use the window optimizer.