Refactor dataset cataloging module
Replace Markdown with JSON, which is much easier to deal with in terms of
appending new data.

Load the full catalog for every new dataset, append the dataset, re-sort the
catalog and update both the json and HTML catalogs.

Rename dataset cataloging output products to catalog.json and index.html.