Don't run packet selection before drift estimates
When drift estimation is active, the initial packet selection estimates prior to
drift estimates can disturb performance figures such as Max|TE|. Also, they can
influence transients, such as that of EWMA, which could go through a second
transient once drift estimates start. Hence, it seems better to simply skip
the samples that precede drift estimates when drift compensation is active.

The way we accomplish this in this patch requires minimal changes to the
underlying code. We only need to change the internal pointer to the
dataset (self.data) such that it starts when drift estimates are available. The
original dataset pointer is also kept and self.data is reset to it on every call
to the process method.