Improve organization of packet selection state
Move the state reset logic into its own function and call it from process every
time, such that user/caller of packet selection does not need to worry about
it. In the previous version, the state variables would only be reset if the
user/caller also called `set_window_len`, which seems to be wrong.