Reorganize reading of serial sync/async data
The main data that we read are the timestamp sets. Anything that is not a
timestamp set is deemed as asynchronous and complementary data. In the new
organization, asynchronous data are stored into queues and then included in the
output dataset asynchronously. The only exception is the temperature, due to its
rate being higher than timestamp sets.

This patch also moves all the reading logic into functions that are reused
between BBU/RRU/RRU2 where possible, such as the occupancy reading, which is
common to all of them.