Configure the PTP message rate based on metadata
The metadata supplied for the acquisition contains the "Sync period", which in
reality corresponds to the period of the two-way exchange that we capture,
namely the PDelayReq/Resp exchange. Previously this metadata field was only
informative. After this patch, however, it will be used to actually configure
the message period (link delay measurement period) adopted by the RRUs.