Add option to load simulation data on analyzer
On the analyze script, run either the testbed dataset reader or the simulation
runner to load the simulated dataset, depending on the type of dataset. Detect
the type of dataset (simulation or testbed) based on the dataset file name.

When running the analysis based on simulated data, skip some plots and analysis
concerning testbed-specific data.