Fix testbed acquisition termination routine
When the acquisition completed the target number of samples, it was going
through a separate termination routine than the one executed upon a signal
handler. For example, when completing based on the number of samples, the
routine was missing the call to "self.roe.serial.stop()".

This patch moves the termination routine to a single function, which is called
by both the interrupt signal handler and the normal termination upon completion
of the target number of samples. It also fixes the moving of the dataset to the
repository in non-interactive mode when catching a signal.