Fix unit testing of vectorized sample-max
We didn't notice the problem because we were previously reusing the same
dictionary on all tests. Test "test_sample_max_vec" was not creating any
"x_pkts_max" results, but the dictionary already had "x_pkts_max" results from a
previous test.

We should start fresh on each test. The way to do so is to create a deep copy of
the original data dictionary.