Refactor the analyzer's metric filtering approach
Instead of explicitly adding a "show_x" argument to each function, simply rely
on a variable-length kwargs processed by the decorator. This approach is easier
to maintain and avoids a lot of code repetition.

This patch also supports more fine-grained control over metrics that could not
be filtered previously. For example, most functions had "show_pkts", but that
was not sufficient to filter a specific packet selection algorithm. It would
only filter out all packet selection algorithms at once. With the kwargs
approach, one can call a plot function with, for instance, "show_pkts_mode" to
omit specifically sample-mode plots.