Add window error plot with all metrics at once
Useful for comparison of the window effect on different estimators.

Change also MSE plots to RMSE as it is easier to visualize on RMSE scale.

For consistency with the other plots, use the labels defined on metrics.py, not
the ones defined on window.py.

Sort window error curves according to the minimum error that each curve
achieves.