Drop ranking fairness approach
Now that the ranking function is reused within the max|TE|/MTIE plot methods, it
becomes to hard to make sure that the max|TE|/MTIE calculated on the ranking
function matches with the one to be calculated by the max|TE|/MTIE plotting
functions. When we use the fairness approach of restricting samples, the range
of indexes changes, so the cached max|TE|/MTIE values differ.

It's not a big deal if we don't rank the algorithms with perfect fairness. At
the moment, we only use this information for quick inspection. Hence, it's
easier to simply drop this feature than to make all the index ranges compatible
between plotting and ranking functions.