Drop "accurate raw" curve on toffset_vs_time plot
We don't really use it that much. This was introduced in the beginning to get a
sense of whether there are accurate measurements within the noisy set of
measurements. If we wanted to preserve this view, we could create a separate
plot containing only the raw measurements and the accurate raw measurements.