Fix unnecessary copying on packet selection window
The original motivation for copying the vector was to be able to reuse the
original ndarray without the changes from the algorithms. However, the current
implementation already creates a new ndarray with time offset measurements or
timestamp differences that we can write in place. So the copying during the
strides trick windowing is unnecessary.