Add recursive sample-min/max implementations
These implementations are a lot faster than the window-based approaches. They
are suitable for sliding window	computations, where one samples enters and one
sample leaves the window on every iteration. The min/max state can be tracked
more efficiently in this case.

Place these implementations on a separate module that can be reused in other
places.

On the PktSelection process function, prefer a recursive implementation when
available. Add option to avoid using the recursive implementation even if
available, so that the other implementations can be tested.

Review the unit tests to cover all sample-min/max implementations.