Merge avg-normal and avg-recursive strategies
Now that the PktSelection.process() method has an option "recursive" to prefer
or disable the preference of a recursive implementation, merge the two
sample-average implementations. If the recursive flag is True (the default), use
the recursive implementation. Otherwise, use the former "normal" window-based
implementation.

Change the keys used to store the results and update the units tests
accordingly.