Reorganize and improve moving-avg implementations
To follow the new packet selection organization for recursive algorithms, move
the recursive moving-average and the EWMA implementation to filters.py
module.

Change the moving-average implementation such that it uses an FIR filter. Given
that the convolution is implemented by low-level code, the result is expected to
be faster, even though it takes more operations to compute the moving average
using a convolution.