Fix cumulative drift estimation error evaluation
When the drift estimation error targets at a given number of
samples (n_samples), the cumulative drift has to be computed before restricting
the subset of instantaneous drift estimation errors. Otherwise, the cumulative
values change. Hence, compute the "true_drift" and "drift_est" vectors with all
samples and wait to restrict the sample range on the resulting instantaneous and
cumulative drift error arrays. Add unit test to cover this computation.