Simplify the TLL transitory skipping strategy
Instead of computing the theoretical settling time for the loop, rely on a
desired maximum settling time specified by the caller. The rationale is that the
analyze.py script defines a portion of the dataset that is skipped on
analysis. This portion of the dataset can be freely used for algorithm
transitory phases. For example, when combining window-based algorithms with
drift compensation, the skipped portion of the dataset is allocated between the
two algorithm stages.

This is an easier and less error-prone approach because it avoids the risk of
the theoretical settling time not matching with the actual settling of the
loop. Given that the loop input is noisy, often with non-Gaussian distributions,
and biased, the theoretical settling time is not guaranteed.

With the approach introduced by this patch, the optimization will already pursue
the best performance within the portion of the dataset that is actually
analyzed (after the transient phase). Hence, this approach can explore even the
configurations that would be ignored in the previous implementation based on the
nominal settling time. That is, even if the nominal settling time says that the
loop won't be settled in the desired portion of the dataset, the new
implementation will still evaluate the loop and check the results.