Fix window optimizer omitting last window length
If we stop at index "i", with the index starting from 0, it means that we
evaluated "i+1" window lengths. When taking the subset of window lengths/errors,
use "x[:(i+1)]" instead of "x[:i]". For convenience, return the index succeeding
the stop index, so that this returned index can be used directly in the
"x[:idx]" notation.