Update initial value for the KF covariance matrix
Initialize the covariance matrix P with the time offset variance R. The
objective is to make the Kalman gain big enough to start the filter
with low confidence on the state. Note that the matrix R corresponds to the
variance of the estimated delay experienced by the PTP messages, so it
will change depending on the FH traffic. In order to not use some
arbitrary big number, we can initialize the matrix P directly with the
variance of the time offset measurement. For the frequency offset, just
use a arbitrary big number.