Refactor batch.py's argument processing scheme
Keep all arguments on a single flattened dictionary. For each process on the
batch of processes, combine the process-specific arguments with the
global (common) arguments.

Let the process-specific arguments override the global arguments. This
overriding is useful to reduce the size of the JSON recipes. We can set a common
parameter globally and override it locally where necessary.

When running each process, convert the dictionary of arguments into a list of
arguments, where each argument key is preceded with `-` or `--`.

Add a workaround for the verbosity argument, which is inherited directly from
batch.py's ArgumentParser object. This argument is counted, and it is an
exception to the general rule for deciding between `-` or `--` prefix. The `-`
prefix applies to a single-character argument key, whereas `--` applies to
multi-character keys. The counted argument uses `-` and is a multi-character
argument. As a workaround, infer that the argument is counted if it is a tuple
on the argument dictionary.