Fix re-computing of max|TE|/MTIE ranking results
The ranking computation happens when calling the ranking function
explicitly (rank_algorithms) or when calling the max|TE|/MTIE plotting
methods. Nevertheless, it is possible that other time offset estimation
results are added later to the dataset, after calling these functions.
Therefore, this patch moves the logic to check if we need to re-compute
the ranking to the ranking function (rank_algorithms).  It also
re-compute the ranking after loading the max|TE|/MTIE results from a
cache file. Thus, this function should be called every time regardless.