Include bias post-compensation on window optimizer
The optimizer must observe the same performance as observed with the final
processing on analyze.py using the optimal window. If bias correction is enabled
there (on analyze.py), it must be enabled on window.py too.

The bias pre-compensation stage was already taking effect on the window
optimizer. In contrast, the post-compensation stage was missing, as it happens
after the call to the window optimizer. This patch adds this missing stage.